requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
openpyxl>=3.1.0
schedule>=1.2.0
gspread>=6.0.0
google-auth>=2.20.0
streamlit>=1.30.0
